Fix build/tests/style
