+++ src/lib.rs
@@ -2,6 +2,7 @@
 #![feature(asm)]
 #![feature(const_fn)]
 #![feature(integer_atomics)]
+#![feature(is_sorted)]
 #![feature(drain_filter)]
 #![feature(nll)]
 #![feature(box_syntax)]
@@ -12,7 +13,6 @@
 #![feature(associated_type_defaults)]
 #![feature(specialization)]
 #![feature(trait_alias)]
-
 // TODO: We should fix missing docs for public items and turn this on (Issue #309).
 // #![deny(missing_docs)]
 
@@ -54,16 +54,18 @@ extern crate num_cpus;
 extern crate downcast_rs;
 
 mod mmtk;
-pub use crate::mmtk::MMTK;
+pub use mmtk::MMTK;
+pub(crate) use mmtk::VM_MAP;
 
 mod policy;
 
 pub mod memory_manager;
-pub mod scheduler;
 pub mod plan;
+pub mod scheduler;
 pub mod util;
 pub mod vm;
 
 pub use crate::plan::{
-    AllocationSemantics, BarrierSelector, CopyContext, Mutator, MutatorContext, Plan, TraceLocal, TransitiveClosure,
+    AllocationSemantics, BarrierSelector, CopyContext, Mutator, MutatorContext, Plan, TraceLocal,
+    TransitiveClosure,
 };
