+++ src/policy/copyspace.rs
@@ -153,6 +153,7 @@ impl<VM: VMBinding> CopySpace<VM> {
         }
     }
 
+    #[allow(dead_code)] // Only used with certain features (such as sanity)
     pub fn protect(&self) {
         if !self.common().contiguous {
             panic!(
@@ -167,6 +168,7 @@ impl<VM: VMBinding> CopySpace<VM> {
         trace!("Protect {:x} {:x}", start, start + extent);
     }
 
+    #[allow(dead_code)] // Only used with certain features (such as sanity)
     pub fn unprotect(&self) {
         if !self.common().contiguous {
             panic!(
