+++ src/policy/mod.rs
@@ -18,5 +18,3 @@ pub mod immortalspace;
 pub mod largeobjectspace;
 pub mod lockfreeimmortalspace;
 pub mod mallocspace;
-
-pub const NUMBER_OF_POLICIES_REQUIRING_SIDE_METADATA: usize = 0;
