+++ src/util/alloc/mod.rs
@@ -1,6 +1,6 @@
 pub(crate) mod allocator;
-pub use allocator::Allocator;
 pub use allocator::fill_alignment_gap;
+pub use allocator::Allocator;
 
 pub(crate) mod allocators;
 pub use allocators::AllocatorSelector;
