+++ src/util/heap/layout/byte_map_mmapper.rs
@@ -18,10 +18,11 @@ const UNMAPPED: u8 = 0;
 const MAPPED: u8 = 1;
 const PROTECTED: u8 = 2;
 
-const MMAP_NUM_CHUNKS: usize = if
-    LOG_BYTES_IN_ADDRESS_SPACE == 32 {
-    1 << (LOG_BYTES_IN_ADDRESS_SPACE as usize - LOG_MMAP_CHUNK_BYTES) }
-    else { 1 << (33 - LOG_MMAP_CHUNK_BYTES) };
+const MMAP_NUM_CHUNKS: usize = if LOG_BYTES_IN_ADDRESS_SPACE == 32 {
+    1 << (LOG_BYTES_IN_ADDRESS_SPACE as usize - LOG_MMAP_CHUNK_BYTES)
+} else {
+    1 << (33 - LOG_MMAP_CHUNK_BYTES)
+};
 pub const VERBOSE: bool = true;
 
 pub struct ByteMapMmapper {
